+++ R/googleAuthR.R
@@ -8,6 +8,15 @@
 #'   \item \code{\link{gar_api_generator}} A function factory for easy enabling of Google API functions.
 #' }
 #' 
+#' @section Batching:
+#' 
+#' If you have many API calls, you can save a lot of time by using batching.  
+#'   This takes your many calls and sends them in one POST request to \code{www.googleapis.com/batch}, 
+#'   returning a list of any responses.  See \code{\link{gar_batch}} for details.
+#'   
+#' Another common batch task is to call the same function with one parameter changing each call.  
+#'  This is supported using the \code{\link{gar_batch_walk}} function.
+#' 
 #' @section Shiny functions:
 #' 
 #' If you need Shiny authentication, then these functions work together to give a smooth authentication flow.